src/verification-service/requirements.txt
fastapi==0.115.8
pydantic==2.10.6
Requests==2.32.3
uvicorn==0.34.0
aio-pika==9.5.5
pydantic-ai-slim[openai, evals, mcp]==0.1.8